# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Bookstore API Documentation",
        "description": "Dokumentasi API untuk Bookstore dengan Rating & Books endpoint",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
        "/api/authors": {
            "get": {
                "tags": [
                    "Authors"
                ],
                "summary": "Get Top Authors",
                "description": "Returns the top 10 authors based on ratings greater than 5.",
                "operationId": "d584d7d2f35059a20073a0b405081cc7",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "name": {
                                                "type": "string",
                                                "example": "Author Name"
                                            },
                                            "top_rating_voter_count": {
                                                "type": "integer",
                                                "example": 12
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get list of books",
                "description": "Returns a paginated list of books with author, category, ratings count and average score.",
                "operationId": "b3edd7a021a4cbcb4fb1e9e98ceaba21",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Filter books by title or author name",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "example": "Harry"
                        }
                    },
                    {
                        "name": "voter_count",
                        "in": "query",
                        "description": "Minimum number of ratings",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 5
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Number of results per page (max 100)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Book Title"
                                                    },
                                                    "author": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Author Name"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "category": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Category Name"
                                                            }
                                                        },
                                                        "type": "object"
                                                    },
                                                    "voter_count": {
                                                        "type": "integer",
                                                        "example": 10
                                                    },
                                                    "average_rating": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": 8.5
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "links": {
                                            "type": "object"
                                        },
                                        "meta": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/books-authors": {
            "get": {
                "tags": [
                    "Ratings"
                ],
                "summary": "List books with their authors (pagination & search)",
                "description": "Retrieve paginated list of books with their authors, with optional search and perPage.",
                "operationId": "ba089617c5388ebe15d7a8d887cfa3bf",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search by book title or author name",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Harry Potter"
                    },
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Number of results per page",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Book Title"
                                                    },
                                                    "author": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "Author Name"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 100
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/ratings": {
            "post": {
                "tags": [
                    "Ratings"
                ],
                "summary": "Submit a rating for a book",
                "description": "Add a rating to a book",
                "operationId": "a69a0d0b66a5422952f805a0533f97d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "book_id",
                                    "score"
                                ],
                                "properties": {
                                    "book_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "score": {
                                        "description": "Rating score between 1-10",
                                        "type": "integer",
                                        "example": 8
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rating submitted successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Rating submitted successfully"
                                        },
                                        "rating": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "book_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "score": {
                                                    "type": "integer",
                                                    "example": 8
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "tags": [
        {
            "name": "Authors",
            "description": "Authors"
        },
        {
            "name": "Books",
            "description": "Books"
        },
        {
            "name": "Ratings",
            "description": "Ratings"
        }
    ]
}